adding some files for a refactor for the validate.py script for future work
